refactor: 모델 파일 분리 (#2)
