+++ model/deep_rec_model.py
@@ -1,0 +1,17 @@
+
+import torch.nn as nn
+
+class DeepRecModel(nn.Module):
+    def __init__(self, input_dim=144):
+        super().__init__()
+        self.mlp = nn.Sequential(
+            nn.Linear(input_dim, 128),
+            nn.ReLU(),
+            nn.Linear(128, 64),
+            nn.ReLU(),
+            nn.Linear(64, 1),
+            nn.Sigmoid()
+        )
+
+    def forward(self, x):
+        return self.mlp(x)
